conversation/it-service-desk-automation/agents/ticket/tools/function/requirements.txt
functions-framework==3.*
google-cloud-firestore
google-cloud-aiplatform